Reorganize utils into separate files
