+++ talelio_backend/app_user/use_cases/authenticate_user.py
@@ -3,7 +3,7 @@ from talelio_backend.core.exceptions import AccountError, TokenError
 from talelio_backend.data.uow import UnitOfWork
 from talelio_backend.identity_and_access.authentication import Authentication
 from talelio_backend.identity_and_access.token_store import TokenStore
-from talelio_backend.shared.utils import generate_time_from_now
+from talelio_backend.shared.utils.time import generate_time_from_now
 
 
 def generate_access_token(user_id: int, username: str) -> str:
