+++ talelio_backend/shared/utils/time.py
@@ -1,0 +1,5 @@
+from datetime import datetime, timedelta
+
+
+def generate_time_from_now(seconds: int) -> datetime:
+    return datetime.utcnow() + timedelta(seconds=seconds)
