+++ talelio_backend/tests/integration/test_user_services.py
@@ -7,7 +7,7 @@ from talelio_backend.app_user.use_cases.authenticate_user import (delete_refresh
                                                                   verify_refresh_token)
 from talelio_backend.core.exceptions import AccountError, TokenError
 from talelio_backend.identity_and_access.authentication import Authentication
-from talelio_backend.shared.utils import generate_time_from_now
+from talelio_backend.shared.utils.time import generate_time_from_now
 from talelio_backend.tests.constants import FAKE_TOKEN, INITIAL_USER_ID, USERNAME_TALEL
 from talelio_backend.tests.integration.helpers import (get_access_token_helper,
                                                        register_account_helper)
@@ -24,11 +24,11 @@ def test_can_get_access_token_for_user() -> None:
 
 
 def test_access_token_expires_after_30_min() -> None:
-    thirtyone_mins_from_now = generate_time_from_now(1860)
+    thirty_one_mins_from_now = generate_time_from_now(1860)
     uow = register_account_helper()
     access_token = get_access_token_helper(uow)
 
-    with freeze_time(thirtyone_mins_from_now):
+    with freeze_time(thirty_one_mins_from_now):
         with pytest.raises(ExpiredSignatureError):
             Authentication.verify_token(access_token)
